+++ src/permission/dto/create-permission.dto.ts
@@ -1,7 +1,8 @@
 import { IsNotEmpty, IsString, Length, Validate } from 'class-validator';
 import { MAX_STRING_LEN, MIN_STRING_LEN } from 'constant';
-import { IsLocalhostUrl } from 'src/utils/url.utils';
+
 import { Url } from 'url';
+import { IsLocalhostUrl } from 'utils';
 
 export class CreatePermissionDto {
   @IsNotEmpty()
