+++ src/role/entity/role.entity.ts
@@ -1,8 +1,8 @@
 import { Roles } from 'src/auth/roles.decorator';
 import { Column, Entity, JoinTable, ManyToMany, OneToMany, PrimaryGeneratedColumn } from "typeorm";
 import { User } from "src/user/entity/user.entity";
-import { CommonColumns } from "src/utils/common.utils";
 import { Permission } from 'src/permission/entity/permission.entity';
+import { CommonColumns } from 'utils';
 
 
 @Entity({'name':'role'})
