+++ src/role/role.module.ts
@@ -1,5 +1,4 @@
 import { PermissionModule } from './../permission/permission.module';
-import { UtilsModule } from 'src/utils/utils.module';
 import { Module, forwardRef } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Role } from './entity/role.entity';
@@ -10,7 +9,6 @@ import { Permission } from 'src/permission/entity/permission.entity';
 @Module({
   imports: [
     TypeOrmModule.forFeature([Role, Permission]),
-    UtilsModule,
     forwardRef(() => PermissionModule),
   ],
   controllers: [RoleController],
